requirements.txt
requests
beautifulsoup4
pandas
camelot-py
openpyxl
fastparquet
duckdb
# opencv-contrib-python
opencv-python-headless